% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back_trans.R
\name{back_trans}
\alias{back_trans}
\title{Back transformation of pooled parameters and calculating
confidence intervals}
\usage{
back_trans(est, se, t)
}
\arguments{
\item{est}{A natural logarithmic transformed pooled parameter
estimate}

\item{se}{A natural logarithmic transformed pooled
standard error}

\item{t}{t-value}
}
\value{
Confidence interval.
}
\description{
\code{back_trans} Back transformation to original scale of
natural logarithmic transformed parameters and calculating
confidence interval
}
\details{
Back transformation of naturally log transformed
and pooled parameter estimates.
}
\author{
Martijn Heymans, 2021
}
