% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_test.R
\name{bf_test}
\alias{bf_test}
\title{Calculates the Brown-Forsythe test.}
\usage{
bf_test(y, x, formula, data)
}
\arguments{
\item{y}{numeric response variable.}

\item{x}{categorical variable.}

\item{formula}{A formula object to specify the model as normally
used by glm. Use 'factor' to define the grouping variable.}
}
\value{
An object containing the following objects:
\itemize{
\item  \code{fstats} F-test value, including numerator and
denominator degrees of freedom.
\item  \code{qhat} pooled coefficients from fit
\item  \code{vcov} variance-covariance matrix
}
}
\description{
\code{bf_test} Calculates the Brown-Forsythe test.
}
\author{
Martijn Heymans, 2021
}
