% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2fz.R
\name{cor2fz}
\alias{cor2fz}
\title{Fisher z transformation of correlation coefficient}
\usage{
cor2fz(r)
}
\arguments{
\item{r}{value for the correlation coefficient.}
}
\value{
correlation coefficient on z scale.
}
\description{
\code{cor2fz} Fisher z transformation of correlation coefficient
}
\examples{
 cor2fz(r=0.65)

}
\author{
Martijn Heymans, 2022
}
