% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2milist.R
\name{df2milist}
\alias{df2milist}
\title{Turns a data frame with multiply imputed data into
an object of class 'milist' to be further used
by 'miceafter::with'}
\usage{
df2milist(data, impvar, keep = FALSE)
}
\arguments{
\item{data}{an object of class 'data.frame', 'tbl_df' or 'tbl'
(tibble).}

\item{impvar}{A character vector. Name of the variable that
distinguishes the imputed datasets.}

\item{keep}{if TRUE the grouping column is kept, if FALSE (default)
the grouping column is not kept.}
}
\value{
an object of class 'milist' (Multiply Imputed Data list)
}
\description{
\code{df2milist} Turns a data frame of class 'data.frame', 'tbl_df'
or 'tbl' (tibble) into an object of class 'milist' to be further used
by 'miceafter::with'
}
\author{
Martijn Heymans, 2021
}
