% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2chi.R
\name{f2chi}
\alias{f2chi}
\title{Converts F-values into Chi Square values}
\usage{
f2chi(f, df_num)
}
\arguments{
\item{f}{a vector of F values.}

\item{df_num}{single value for the numerator degrees of freedom of the F test.}
}
\value{
The Chi square values.
}
\description{
\code{f2chi} convert F to Chi-square values.
}
\examples{
  f2chi(c(5.83, 4.95, 3.24, 6.27, 4.81), 5)

}
\author{
Martijn Heymans, 2021
}
