% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fz2cor.R
\name{fz2cor}
\alias{fz2cor}
\title{Fisher z back transformation of correlation coefficient}
\usage{
fz2cor(z)
}
\arguments{
\item{z}{value of the correlation coefficient on z scale.}
}
\value{
correlation coefficient on correlation scale.
}
\description{
\code{fz2cor} Fisher z back transformation of correlation coefficient
}
\examples{
 fz2cor(z=0.631)

}
\author{
Martijn Heymans, 2022
}
