% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit.R
\name{inv_logit}
\alias{inv_logit}
\title{Takes the inverse of logit transformed parameters and calculates
the confidence intervals}
\usage{
inv_logit(est, se, crit.value)
}
\arguments{
\item{est}{A parameter estimate on the logit scale.}

\item{se}{A standard error value on the logit scale.}

\item{crit.value}{Critical value of any distribution.}
}
\value{
Parameter, critical value and confidence
intervals on original scale.
}
\description{
\code{inv_logit} Takes the inverse of logit transformed
parameters and calculates the confidence interval
by using the critical value.
}
\details{
Takes the inverse of logit transformed parameter
estimates. The confidence interval is calculated by taking the
inverse of \eqn{est +/- crit.value{1-\alpha/2} * se}.
}
\examples{
 inv_logit(est=1.39, se=0.25, crit.value=1.96)

}
\author{
Martijn Heymans, 2021
}
