% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invlogit.R
\name{invlogit}
\alias{invlogit}
\title{Takes the inverse of a logit transformed value}
\usage{
invlogit(est)
}
\arguments{
\item{est}{A parameter estimate on the logit scale.}
}
\value{
back transformed value.
}
\description{
\code{invlogit} Takes the inverse of a logit transformed
value
}
\examples{
 invlogit(est=1.39)

}
\author{
Martijn Heymans, 2021
}
