% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbp_orig.R
\docType{data}
\name{lbp_orig}
\alias{lbp_orig}
\title{Dataset of 159 Low Back Pain Patients with missing values}
\format{
A data frame with 159 observations on the following 15 variables.
\describe{
\item{Chronic}{dichotomous}
\item{Gender}{dichotomous}
\item{Carrying}{categorical}
\item{Pain}{continuous}
\item{Tampascale}{continuous}
\item{Function}{continuous}
\item{Radiation}{dichotomous}
\item{Age}{continuous}
\item{Smoking}{dichotomous}
\item{Satisfaction}{categorical}
\item{JobControl}{continuous}
\item{JobDemands}{continuous}
\item{SocialSupport}{continuous}
\item{Duration}{continuous}
\item{BMI}{continuous}
}
}
\usage{
lbp_orig
}
\description{
A data frame with 159 observations of 15 variables related to low back pain.
}
\examples{
 data(lbp_orig)
 ## maybe str(lbp_orig)

}
\keyword{datasets}
