% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbpmicox.R
\docType{data}
\name{lbpmicox}
\alias{lbpmicox}
\title{Survival data of 265 Low Back Pain Patients}
\format{
A data frame with 2650 observations on the following 18 variables.
\describe{
\item{Impnr}{a numeric vector}
\item{patnr}{a numeric vector}
\item{Status}{dichotomous event}
\item{Time}{continuous follow up time variable}
\item{Duration}{continuous}
\item{Previous}{dichotomous}
\item{Radiation}{dichotomous}
\item{Onset}{dichotomous}
\item{Age}{continuous}
\item{Tampascale}{continuous}
\item{Pain}{continuous}
\item{Function}{continuous}
\item{Satisfaction}{categorical}
\item{JobControl}{continuous}
\item{JobDemand}{continuous}
\item{Social}{continuous}
\item{Expectation}{a numeric vector}
\item{Expect_cat}{categorical}
}
}
\usage{
lbpmicox
}
\description{
A data frame with 10 multiply imputed datasets of 265 observations each on
17 variables related to low back pain.
}
\examples{
 data(lbpmicox)
 ## maybe str(lbpmicox)

}
\keyword{datasets}
