% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbpmilr.R
\docType{data}
\name{lbpmilr}
\alias{lbpmilr}
\title{Data of 159 Low Back Pain Patients}
\format{
A data frame with 1590 observations on the following 17 variables.
\describe{
\item{Impnr}{a numeric vector}
\item{ID}{a numeric vector}
\item{Chronic}{dichotomous}
\item{Gender}{dichotomous}
\item{Carrying}{categorical}
\item{Pain}{continuous}
\item{Tampascale}{continuous}
\item{Function}{continuous}
\item{Radiation}{dichotomous}
\item{Age}{continuous}
\item{Smoking}{dichotomous}
\item{Satisfaction}{categorical}
\item{JobControl}{continuous}
\item{JobDemands}{continuous}
\item{SocialSupport}{continuous}
\item{Duration}{continuous}
\item{BMI}{continuous}
}
}
\usage{
lbpmilr
}
\description{
A data frame with 10 multiply imputed datasets of 159 observations each on
17 variables related to low back pain.
}
\examples{
 data(lbpmilr)
 ## maybe str(lbpmilr)

}
\keyword{datasets}
