% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levene_test.R
\name{levene_test}
\alias{levene_test}
\title{Calculates the Levene's test}
\usage{
levene_test(y, x, formula, data)
}
\arguments{
\item{y}{numeric (continuous) response variable.}

\item{x}{categorical groups variable.}

\item{formula}{A formula object to specify the model as normally
used by glm. Use 'factor' to define the grouping x variable.}

\item{data}{An objects of class \code{mids}, created by
\code{make_mids} or after a call to function \code{mice}.}
}
\value{
An object from which the following objects can be extracted:
\itemize{
\item  \code{fstats} F-test value, including numerator and
denominator degrees of freedom.
\item  \code{qhat} model coefficients.
\item  \code{vcov} variance-covariance matrix.
\item  \code{dfcom} degrees of freedom obtained from /code{df.residual}.
}
}
\description{
\code{levene_test} Calculates the Levene's test for homogeneity
of variance across groups, coefficients, variance-covariance matrix,
and degrees of freedom to be used in function \code{with.miceafter}.
}
\examples{
imp_dat <- make_mids(lbpmilr, impvar="Impnr")
ra <- with.miceafter(imp_dat, expr=levene_test(Pain ~ factor(Carrying)))

}
\seealso{
\code{\link{with.miceafter}}
}
\author{
Martijn Heymans, 2021
}
