% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2milist.R
\name{list2milist}
\alias{list2milist}
\title{Turns a list object with multiply imputed datasets into an object
of class 'milist'.}
\usage{
list2milist(data)
}
\arguments{
\item{data}{an object of class 'list'.}
}
\value{
an object of class 'milist'
}
\description{
\code{list2milist} Turns a list with multiply imputed datasets
into an object of class 'milist' to be further used by 'with.milist'
}
\author{
Martijn Heymans, 2021
}
