% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_trans.R
\name{log_trans}
\alias{log_trans}
\title{Natural Log transformation of parameter estimates}
\usage{
log_trans(est, se)
}
\arguments{
\item{est}{A numeric vector of values that have to be log transformed}

\item{se}{A numeric vector of standard error values}
}
\value{
The natural log transformed values.
}
\description{
\code{log_transform} Natural Log transformation of parameter
estimate and standard error.
}
\details{
Function is used to naturally log transform parameters
and standard errors before they can be pooled using function
\code{pool_RR} before pooling with Rubin's Rules. To naturally
log transform the standard error the Delta method is used.
}
\author{
Martijn Heymans, 2021
}
