% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_trans.R
\name{logit_trans}
\alias{logit_trans}
\title{Logit transformation of parameter estimates}
\usage{
logit_trans(est, se)
}
\arguments{
\item{est}{A numeric vector of values.}

\item{se}{A numeric vector of standard error values.}
}
\value{
The logit transformed values.
}
\description{
\code{logit_trans} Logit transformation of parameter
estimate and standard error.
}
\details{
Function is used to logit transform parameters
and standard errors. For the standard error the
Delta method is used.
}
\author{
Martijn Heymans, 2021
}
