% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mids.R
\name{make_mids}
\alias{make_mids}
\title{Converts a data.frame into a object of class mids}
\usage{
make_mids(data, impvar = NULL)
}
\arguments{
\item{data}{Data frame with stacked multiple imputed datasets.
The original dataset that contains missing values must be excluded from the
dataset. The imputed datasets must be distinguished by an imputation variable,
specified under impvar, and starting by 1.}

\item{impvar}{A character vector. Name of the variable that distinguishes the
imputed datasets.}
}
\value{
Object of class mids.
}
\description{
\code{make_mids} Converta data.frame into object of class mids
}
\author{
Martijn Heymans, 2021
}
