% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mids2milist.R
\name{mids2milist}
\alias{mids2milist}
\title{Turns a 'mice::mids' object into an object of class 'milist'
to be further used by 'miceafter::with'}
\usage{
mids2milist(data, keep = FALSE)
}
\arguments{
\item{data}{a 'mice::mids' object}

\item{keep}{if TRUE the grouping column is kept, if FALSE
(default) the grouping column is not kept.}
}
\value{
an object of class 'milist'
}
\description{
\code{mids2milist} Turns a 'mice::mids' object into an object
of class 'milist' to be further used by 'miceafter::with'
}
\author{
Martijn Heymans, 2021
}
