% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_bftest.R
\name{pool_bftest}
\alias{pool_bftest}
\title{Calculates the pooled Brown-Forsythe test.}
\usage{
pool_bftest(object, method = "D1")
}
\arguments{
\item{object}{An object of class 'mistats'
('Multiply Imputed Statistical Analysis').}

\item{method}{A character vector to choose the pooling method,
'D1' (default) or 'D2'.}
}
\value{
The (combined) F-statistic, p-value and degrees of freedom.
}
\description{
\code{pool_levenetest} Calculates the pooled F-statistic
of the Brown-Forsythe test.
}
\examples{

imp_dat <- df2milist(lbpmilr, impvar="Impnr")
ra <- with(imp_dat, expr=bf_test(Pain ~ factor(Carrying)))
res <- pool_bftest(ra)
res

}
\references{
Eekhout I, van de Wiel MA, Heymans MW. Methods for significance testing of categorical
covariates in logistic regression models after multiple imputation: power and applicability
analysis. BMC Med Res Methodol. 2017;17(1):129.

Enders CK (2010). Applied missing data analysis. New York: The Guilford Press.

Van Buuren S. (2018). Flexible Imputation of Missing Data. 2nd Edition. Chapman & Hall/CRC
Interdisciplinary Statistics. Boca Raton.
}
\seealso{
\code{\link{with.milist}}, \code{\link{bf_test}}
}
\author{
Martijn Heymans, 2021
}
