% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_cindex.R
\name{pool_cindex}
\alias{pool_cindex}
\title{Calculates the pooled C-index and Confidence intervals}
\usage{
pool_cindex(data, conf.level = 0.95, dfcom = NULL)
}
\arguments{
\item{data}{An object of class 'raami' (repeated analysis after
multiple imputation) or a 2 x m matrix with C-index
values and standard errors. For the latter situation dfcom has to be
provided.}

\item{conf.level}{conf.level Confidence level of the confidence intervals.}

\item{dfcom}{Number of completed-data analysis degrees of freedom.}
}
\value{
The pooled c-index value and the confidence intervals.
}
\description{
\code{pool_cindex} Calculates the pooled C-index and Confidence intervals.
}
\details{
Rubin's Rules are used for pooling. The C-index values are log
transformed before pooling and finally back transformed.
}
\examples{
 # Logistic Regression
 imp_dat <- df2milist(lbpmilr, impvar="Impnr")
 ra <- with(data=imp_data,
  expr = cindex(glm(Chronic ~ Gender + Radiation,
  family=binomial)))
 res <- pool_cindex(data)
 res

 # Cox regression
 imp_dat <- df2milist(lbpmilr, impvar="Impnr")
  ra <- with(data=imp_data,
  expr = cindex(coxph(Surv(Time, Status) ~ Pain + Radiation)))
 res <- pool_cindex(data)
 res

}
\seealso{
\code{\link{with.milist}}, \code{\link{cindex}}
}
\author{
Martijn Heymans, 2021
}
