% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_oddsratio.R
\name{pool_oddsratio}
\alias{pool_oddsratio}
\title{Calculates the pooled odds ratio (OR) and related confidence interval.}
\usage{
pool_oddsratio(object, conf.level = 0.95)
}
\arguments{
\item{object}{An object of class 'raami' (repeated analysis after
multiple imputation) after using \code{with.miceafter}.}

\item{conf.level}{Confidence level of the confidence intervals.}
}
\value{
The pooled OR and confidence intervals.
}
\description{
\code{pool_oddsratio} Calculates the pooled odds ratio and
confidence interval.
}
\examples{
imp_dat <- make_mids(lbpmilr, impvar="Impnr")
ra <- with.miceafter(imp_dat, expr=oddsratio(Chronic ~ Radiation))
res <- pool_oddsratio(ra)
res

}
\author{
Martijn Heymans, 2021
}
