% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_prop_wilson.R
\name{pool_prop_wilson}
\alias{pool_prop_wilson}
\title{Calculates the pooled single proportion confidence intervals according
to Wilson across multiply imputed datasets.}
\usage{
pool_prop_wilson(object, conf.level = 0.95)
}
\arguments{
\item{object}{An object of class 'staami' after using \code{with.aftermi}.}

\item{formula}{A formula object to specify the model as normally used by glm.}

\item{data}{Data frame with stacked multiple imputed datasets.
The original dataset that contains missing values must be excluded from the
dataset. The imputed datasets must be distinguished by an imputation variable,
specified under impvar, and starting by 1.}
}
\value{
The proportion and the 95\% Confidence interval according to Wilson.
}
\description{
\code{pool_prop_wilson} Calculates the pooled single proportion and
confidence intervals according to Wald across multiply imputed datasets
and using Rubin's Rules.
}
\details{
Before pooling, the proportions will be naturally log transformed and
the pooled estimates back transformed to the original scale.
}
\author{
Martijn Heymans, 2021
}
