% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_propdiff_ac.R
\name{pool_propdiff_ac}
\alias{pool_propdiff_ac}
\title{Calculates the pooled difference between proportions and standard error
according to Agresti-Caffo across multiply imputed datasets.}
\usage{
pool_propdiff_ac(object, conf.level = 0.95)
}
\arguments{
\item{object}{An object of class 'raami' (repeated analysis after
multiple imputation) after using \code{with.aftermi}.}

\item{conf.level}{Confidence level of the confidence intervals.}
}
\value{
The proportion, the Confidence intervals, the standard error and statistic.
}
\description{
\code{pool_prop_wald} Calculates the pooled difference between proportions
and standard error according to Agresti-Caffo across multiply imputed datasets.
}
\examples{
imp_dat <- make_mids(lbpmilr, impvar="Impnr")
ra <- with.miceafter(imp_dat, expr=propdiff_ac(Chronic ~ Gender))
res <- pool_propdiff_ac(ra)
res

}
\author{
Martijn Heymans, 2021
}
