% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_propdiff_wilson.R
\name{pool_propdiff_wilson}
\alias{pool_propdiff_wilson}
\title{Calculates the pooled single proportion confidence intervals according
to Wilson across multiply imputed datasets.}
\usage{
pool_propdiff_wilson(object, conf.level = 0.95)
}
\arguments{
\item{object}{An object of class 'mistats' (repeated statistical
analysis across multiply imputed datasets).}

\item{conf.level}{Confidence level of the confidence intervals. Mostly
set at 0.95.}
}
\value{
The Proportion and the Confidence intervals according to Wilson.
}
\description{
\code{pool_prop_wilson} Calculates the pooled single proportion and
confidence intervals according to Wald across multiply imputed datasets
and using Rubin's Rules.
}
\details{
The \code{pool_propdiff_wilson} function uses information from separate
exposure groups. It is therefore important to first use the \code{propdiff_wald}
function and to set strata = TRUE in that function.
}
\examples{

imp_dat <- df2milist(lbpmilr, impvar="Impnr")
res <- with(imp_dat,
 expr=propdiff_wald(Chronic ~ Radiation, strata = TRUE))
res <- pool_propdiff_wilson(res)
res

}
\references{
Yulia Sidi & Ofer Harel (2021): Difference Between Binomial
Proportions Using Newcombe’s Method With Multiple Imputation for Incomplete
Data, The American Statistician, DOI:10.1080/00031305.2021.1898468
}
\seealso{
\code{\link{with.milist}}, \code{\link{propdiff_wald}}
}
\author{
Martijn Heymans, 2021
}
