% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_riskratio.R
\name{pool_riskratio}
\alias{pool_riskratio}
\title{Calculates the pooled risk ratio (RR) and related confidence interval.}
\usage{
pool_riskratio(object, conf.level = 0.95)
}
\arguments{
\item{object}{An object of class 'mistats' ('Multiply Imputed
Statistical Analysis').}

\item{conf.level}{Confidence level of the confidence intervals.}
}
\value{
The pooled RR and confidence intervals.
}
\description{
\code{pool_riskratio} Calculates the pooled risk ratio and
confidence interval.
}
\examples{
imp_dat <- df2milist(lbpmilr, impvar="Impnr")
ra <- with(imp_dat, expr=riskratio(Chronic ~ Radiation))
res <- pool_riskratio(ra)
res

}
\seealso{
\code{\link{with.milist}}, \code{\link{riskratio}}
}
\author{
Martijn Heymans, 2021
}
