% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_diff_wald.R
\name{prop_diff_wald}
\alias{prop_diff_wald}
\title{Calculates the difference between proportions according to Wald}
\usage{
prop_diff_wald(formula, data)
}
\arguments{
\item{formula}{A formula object to specify the model as normally used by glm.}

\item{data}{An objects of class \code{mids}, created by
\code{make_mids} or after a call to function \code{mice}.
If \code{data} is of type \code{data.frame}, use
\code{make_mids} to convert to \code{mids} object.}
}
\value{
The difference between proportions and the standard error.
}
\description{
\code{prop_diff_wald} Calculates the difference between proportions
acoording to Wald to be used in \code{with.aftermi}.
}
\author{
Martijn Heymans, 2021
}
