% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskratio.R
\name{riskratio}
\alias{riskratio}
\title{Calculates the risk ratio and standard error.}
\usage{
riskratio(y, x, formula, data)
}
\arguments{
\item{y}{0-1 binary response variable.}

\item{x}{0-1 binary independent variable.}

\item{formula}{A formula object to specify the model as normally used by glm.}

\item{data}{An objects of class \code{mids}, created by
\code{make_mids} or after a call to function \code{mice}.
If \code{data} is of type \code{data.frame}, use
\code{make_mids} to convert to \code{mids} object.}
}
\value{
The risk ratio and related standard error and degrees of freedom.
}
\description{
\code{riskratio} Calculates the risk ratio and standard error.
}
\author{
Martijn Heymans, 2021
}
