% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.aftermi.R
\name{with.aftermi}
\alias{with.aftermi}
\title{Evaluate an Expression across multiply imputed datasets}
\usage{
\method{with}{aftermi}(data, expr = NULL, impvar = NULL)
}
\arguments{
\item{data}{Data that is used to evaluate the expression in.
Can be an objects of class \code{mids} after a call to function \code{mice},
or to function \code{make_mids}. Or an object of class 'data.frame'
consisting of a dataset containing the stacked imputed datasets (excluding)
the original dataset (normally indicated as dataset 0). The imputed datasets
must be distinguished by an imputation variable, specified under impvar,
and starting by 1.}

\item{expr}{expression to evaluate.}

\item{impvar}{A character vector. Name of the variable that distinguishes the
imputed datasets. Only necessary when data object is a 'data.frame'.}
}
\value{
The value of the evaluated expression with class \code{raami} 'repeated
analyses after multiple imputation'.
}
\description{
\code{with.aftermi} Evaluate an expression in the form of a
statistical test procedure across multiply imputed datasets
}
\author{
Martijn Heymans, 2021
}
