% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.milist.R
\name{with.milist}
\alias{with.milist}
\title{Evaluate an Expression across a list of multiply imputed datasets}
\usage{
\method{with}{milist}(data, expr = NULL, ...)
}
\arguments{
\item{data}{data that is used to evaluate the expression in,
an objects of class \code{milist} after a call to function
\code{df2milist} or \code{mids2milist}. For 'df2milist'
the original dataset (normally indicated as dataset 0) must be
exluded and the imputed datasets must be distinguished by an
imputation variable, specified under impvar and starting by 1.}

\item{expr}{expression to evaluate.}
}
\value{
The value of the evaluated expression with class \code{mistats}
'statistical analyses in multiply imputed data'.
}
\description{
\code{with.milist} Evaluate an expression in the form of a
statistical test procedure across a list of multiply imputed datasets
}
\author{
Martijn Heymans, 2021
}
