% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Function to check input data for function \code{psfmi_lr}}
\usage{
check_model(
  data,
  formula,
  keep.predictors,
  impvar,
  p.crit,
  method,
  nimp,
  direction,
  model_type
)
}
\arguments{
\item{data}{Data frame with stacked multiple imputed datasets.
The original dataset that contains missing values must be excluded from the
dataset. The imputed datasets must be distinguished by an imputation variable,
specified under impvar, and starting by 1.}

\item{formula}{A formula object to specify the model as normally used by glm.
See under "Details" and "Examples" how these can be specified. If a formula
object is used set predictors, cat.predictors, spline.predictors or int.predictors
at the default value of NULL.}

\item{keep.predictors}{A single string or a vector of strings including the variables that are forced
in the model during predictor selection. All type of variables are allowed.}

\item{impvar}{A character vector. Name of the variable that distinguishes the
imputed datasets.}

\item{p.crit}{A numerical scalar. P-value selection criterium. A value of 1
provides the pooled model without selection.}

\item{method}{A character vector to indicate the pooling method for p-values to pool the
total model or used during predictor selection. This can be "RR", D1", "D2", "D3", "D4", or "MPR".
See details for more information. Default is "RR".}

\item{nimp}{A numerical scalar. Number of imputed datasets. Default is 5.}

\item{direction}{The direction of predictor selection, "BW" means backward selection and "FW"
means forward selection.}

\item{model_type}{A character vector for type of model, "binomial" is for logistic regression and
"linear" is for linear regression models.}
}
\description{
\code{check_model} Back transformation to original scale of
natural logarithmic transformed parameters and calculating
confidence interval
}
\author{
Martijn Heymans, 2020
}
\keyword{internal}
