% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{cindex}
\alias{cindex}
\title{Calculates the c-index and standard error}
\usage{
cindex(formula, data)
}
\arguments{
\item{formula}{A formula object to specify the model as
normally used by glm or coxph.}

\item{data}{An objects of class \code{mids}, created by
\code{make_mids} or after a call to function \code{mice}.
If \code{data} is of type \code{data.frame}, use
\code{make_mids} to convert to a \code{mids} object.}
}
\value{
The c-index, related standard error and
degrees of freedom.
}
\description{
\code{cindex} Calculates the c-index and standard error for
logistic and Cox regression models and the degrees of freedom
to be further used in function \code{with.miceafter}.
}
\examples{

 imp_dat <- df2milist(lbpmilr, impvar="Impnr"
 ra <- with(data=imp_data,
  expr = cindex(glm(Chronic ~ Gender + Radiation,
  family=binomial)))

}
\seealso{
\code{\link{with.milist}}, \code{\link{pool_cindex}}
}
\author{
Martijn Heymans, 2021
}
