% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levene_test.R
\name{levene_test}
\alias{levene_test}
\title{Calculates the Levene's test}
\usage{
levene_test(y, x, formula, data)
}
\arguments{
\item{y}{numeric response variable.}

\item{x}{categorical variable.}

\item{formula}{A formula object to specify the model as normally
used by glm. Use 'factor' to define the grouping variable.}

\item{data}{An objects of class \code{mids}, created by
\code{make_mids} or after a call to function \code{mice}.
If \code{data} is of type \code{data.frame}, use
\code{make_mids} to convert to \code{mids} object.}
}
\value{
An object of class \code{raami} (repeated analyses after
multiple imputation) from which the following objects can be extracted:
\itemize{
\item  \code{fstats} F-test value, including numerator and
denominator degrees of freedom.
\item  \code{qhat} pooled coefficients from fit
\item  \code{vcov} variance-covariance matrix
}
}
\description{
\code{levene_test} Calculates the Levene's test.
}
\author{
Martijn Heymans, 2021
}
