% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mids.R
\name{make_mids}
\alias{make_mids}
\title{Converts a data.frame, tibble or list into a object of class \code{mice::mids}}
\usage{
make_mids(data, impvar = NULL)
}
\arguments{
\item{data}{A Data frame, tibble or list containing the multiply imputed
datasets. The original dataset that contains missing values must be
excluded from the dataset. For data frames and tibbles, The imputed
datasets must be distinguished by an imputation variable,
specified under impvar, and starting by 1.}

\item{impvar}{A character vector. Name of the variable that distinguishes the
imputed datasets.}
}
\value{
Object of class mids.
}
\description{
\code{make_mids} Converts data.frame, tibble or list into object of
class \code{mice::mids}
}
\author{
Martijn Heymans, 2021
}
