% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsratio.R
\name{oddsratio}
\alias{oddsratio}
\title{Calculates the odds ratio (OR) and standard error.}
\usage{
oddsratio(y, x, formula, data)
}
\arguments{
\item{y}{0-1 binary response variable.}

\item{x}{0-1 binary independent variable.}

\item{formula}{A formula object to specify the model as normally used by glm.}

\item{data}{An objects of class \code{milist}, created by
\code{df2milist}, \code{df2milist} or \code{mids2milist}.}
}
\value{
The odds ratio, related standard error and degrees of freedom.
}
\description{
\code{oddsratio} Calculates the odds ratio and standard error
and degrees of freedom to be used in function \code{with.miceafter}.
}
\details{
Note that the standard error of the OR is on the
logit scale.
}
\examples{

imp_dat <- df2milist(lbpmilr, impvar="Impnr")
ra <- with(imp_dat, expr=oddsratio(Chronic ~ Radiation))

}
\seealso{
\code{\link{with.milist}}, \code{\link{pool_oddsratio}}
}
\author{
Martijn Heymans, 2021
}
