% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_oddsratio.R
\name{pool_oddsratio}
\alias{pool_oddsratio}
\title{Calculates the pooled odds ratio (OR) and related confidence interval.}
\usage{
pool_oddsratio(object, conf.level = 0.95)
}
\arguments{
\item{object}{An object of class 'mistats' ('Multiply Imputed
Statistical Analysis')}

\item{conf.level}{Confidence level of the confidence intervals.}
}
\value{
The pooled OR and confidence intervals.
}
\description{
\code{pool_oddsratio} Calculates the pooled odds ratio and
confidence interval.
}
\examples{
imp_dat <- df2milist(lbpmilr, impvar="Impnr")
ra <- with(imp_dat, expr=oddsratio(Chronic ~ Radiation))
res <- pool_oddsratio(ra)
res

}
\seealso{
\code{\link{with.milist}}, \code{\link{oddsratio}}
}
\author{
Martijn Heymans, 2021
}
