% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_prop_nna.R
\name{pool_prop_nna}
\alias{pool_prop_nna}
\title{Calculates the pooled proportion and standard error according
to Wald across multiply imputed datasets.}
\usage{
pool_prop_nna(object, conf.level = 0.95)
}
\arguments{
\item{object}{An object of class 'mistats' (repeated statistical
analysis across multiply imputed datasets).}

\item{conf.level}{Confidence level of the confidence intervals.}
}
\value{
The proportion and the 95\% Confidence interval.
}
\description{
\code{pool_prop_wald} Calculates the pooled proportion and
standard error according to Wald across multiply imputed datasets
and using Rubin's Rules.
}
\details{
Before pooling, the proportions will be naturally log transformed and
the pooled estimates back transformed to the original scale.
}
\examples{
 imp_dat <- df2milist(lbpmilr, impvar='Impnr')
 ra <- with(imp_dat, expr=prop_nna(Radiation))
 res <- pool_prop_nna(ra)
 res

}
\references{
Raghunathan, T. (2016). Missing Data Analysis in Practice.
Boca Raton, FL: Chapman and Hall/CRC. (paragr 4.6.2)
}
\seealso{
\code{\link{with.milist}}, \code{\link{prop_nna}}
}
\author{
Martijn Heymans, 2021
}
