% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_prop_nna.R
\name{pool_prop_nna}
\alias{pool_prop_nna}
\title{Calculates the pooled proportion and standard error according
to Wald across multiply imputed datasets.}
\usage{
pool_prop_nna(object, conf.level = 0.95)
}
\arguments{
\item{object}{An object of class 'raami' after using \code{with.aftermi}.}

\item{data}{Data frame with stacked multiple imputed datasets.
The original dataset that contains missing values must be excluded from the
dataset. The imputed datasets must be distinguished by an imputation variable,
specified under impvar, and starting by 1.}
}
\value{
The proportion and the 95\% Confidence interval.
}
\description{
\code{pool_prop_wald} Calculates the pooled proportion and
standard error according to Wald across multiply imputed datasets
and using Rubin's Rules.
}
\details{
Before pooling, the proportions will be naturally log transformed and
the pooled estimates back transformed to the original scale.
}
\examples{
 lbpmilr <- psfmi::lbpmilr

 imp_dat <- make_mids(lbpmilr, impvar='Impnr')

 ra <- miceafter::with(imp_dat, expr=prop_nna(Radiation))

 pool_prop_na(ra)

}
\references{
Raghunathan, T. (2016). Missing Data Analysis in Practice.
Boca Raton, FL: Chapman and Hall/CRC.
}
\author{
Martijn Heymans, 2021
}
