% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_nna.R
\name{prop_nna}
\alias{prop_nna}
\title{Calculates the proportion and standard error according to
a nonnormal approximation}
\usage{
prop_nna(x, data)
}
\arguments{
\item{x}{name of variable to calculate proportion.}

\item{data}{An objects of class \code{mids}, created by
\code{make_mids} or after a call to function \code{mice}.
If \code{data} is of type \code{data.frame}, use
\code{make_mids} to convert to \code{mids} object.}
}
\value{
The mean and variance components of the beta distribution.
}
\description{
\code{prop_nna} Calculates the proportion and
related confidence interval using a beta distribution.
}
\author{
Martijn Heymans, 2021
}
