% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_wald.R
\name{prop_wald}
\alias{prop_wald}
\title{Calculates a single proportion and related standard error
according to Wald}
\usage{
prop_wald(x, formula, data)
}
\arguments{
\item{x}{name of variable to calculate proportion.}

\item{formula}{A formula object to specify the model as normally used by glm.}

\item{data}{An objects of class \code{mids}, created by
\code{make_mids} or after a call to function \code{mice}.
If \code{data} is of type \code{data.frame}, use
\code{make_mids} to convert to \code{mids} object.}
}
\value{
The proportion and the standard error.
}
\description{
\code{prop_wald} Calculates a single proportion and
related standard error according to Wald and
provides degrees of freedom to be used
in function \code{with.miceafter}.
}
\examples{
imp_dat <- make_mids(lbpmilr, impvar="Impnr")
ra <- with.miceafter(imp_dat, expr=prop_wald(Chronic ~ 1))

}
\seealso{
\code{\link{with.miceafter}}
}
\author{
Martijn Heymans, 2021
}
