% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propdiff_ac.R
\name{propdiff_ac}
\alias{propdiff_ac}
\title{Calculates the difference between proportions and standard error
according to method Agresti-Caffo}
\usage{
propdiff_ac(y, x, formula, data)
}
\arguments{
\item{formula}{A formula object to specify the model as normally used by glm.}

\item{data}{An objects of class \code{mids}, created by
\code{make_mids} or after a call to function \code{mice}.
If \code{data} is of type \code{data.frame}, use
\code{make_mids} to convert to \code{mids} object.}
}
\value{
The Agresti-Caffo difference between proportions and the standard error.
}
\description{
\code{propdiff_ac} Calculates the difference between proportions
and standard error according to method Agresti-Caffo and
degrees of freedom to be used in function \code{with.miceafter}
}
\examples{
imp_dat <- df2milist(lbpmilr, impvar="Impnr")
ra <- with(imp_dat, expr=propdiff_ac(Chronic ~ Radiation + Gender))

}
\seealso{
\code{\link{with.milist}}, \code{\link{pool_propdiff_ac}}
}
\author{
Martijn Heymans, 2021
}
