% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propdiff_wald.R
\name{propdiff_wald}
\alias{propdiff_wald}
\title{Calculates the difference between proportions and standard error
according to Wald}
\usage{
propdiff_wald(y, x, formula, data, strata = FALSE)
}
\arguments{
\item{y}{0-1 binary response variable.}

\item{x}{0-1 binary independent variable.}

\item{formula}{A formula object to specify the model as normally used by glm.}

\item{data}{An objects of class \code{mids}, created by
\code{make_mids} or after a call to function \code{mice}.
If \code{data} is of type \code{data.frame}, use
\code{make_mids} to convert to \code{mids} object.}

\item{strata}{If TRUE the proportion, se and n of each group is provided.
Default is FALSE. Has to be used in combination with function
\code{pool_propdiff_wilson}}
}
\value{
The difference between proportions, standard error and complete
data degrees of freedom.
}
\description{
\code{propdiff_wald} Calculates the difference between proportions and
standard error according to Wald and degrees of freedom to
be used in function \code{with.miceafter}.
}
\examples{
imp_dat <- make_mids(lbpmilr, impvar="Impnr")
ra <- with.miceafter(imp_dat, expr=propdiff_wald(Chronic ~ Gender))

# proportions in each subgroup
imp_dat <- make_mids(lbpmilr, impvar="Impnr")
ra <- with.miceafter(imp_dat, expr=propdiff_wald(Chronic ~ Gender, strata=TRUE))

}
\seealso{
\code{\link{with.miceafter}}, \code{\link{pool_propdiff_wilson}}
}
\author{
Martijn Heymans, 2021
}
